+++ lib/dbReplace.js
@@ -15,7 +15,8 @@ function dbReplace(search,replace,output_file) {
     });
 
     grunt.log.writeln("Replacing '" + search + "' with '" + replace + "' in the database.");
-     // Execute cmd
+
+    // Execute cmd
     shell.exec(cmd);
     grunt.log.oklns("Database references succesfully updated.");
 }
