+++ tasks/deployments.js
@@ -12,9 +12,12 @@
 var shell = require('shelljs');
 
 // Library modules
-var tpls        = require('../lib/tpls');
-var dbReplace   = require('../lib/dbReplace');
-var dbDump      = require('../lib/dbDump');
+var tpls                    = require('../lib/tpls');
+var dbReplace               = require('../lib/dbReplace');
+var dbDump                  = require('../lib/dbDump');
+var dbImport                = require('../lib/dbImport');
+var generateBackupPaths     = require('../lib/generateBackupPaths');
+
 
 // Only Grunt registration within this "exports"
 module.exports = function(grunt) {
@@ -40,8 +43,8 @@ module.exports = function(grunt) {
         var local_options       = grunt.config.get('deployments').local;
 
         // Generate required backup directories and paths
-        var local_backup_paths  = generate_backup_paths("local", task_options);
-        var target_backup_paths = generate_backup_paths(target, task_options);
+        var local_backup_paths  = generateBackupPaths("local", task_options);
+        var target_backup_paths = generateBackupPaths(target, task_options);
 
 
         grunt.log.subhead("Pushing database from 'Local' to '" + target_options.title + "'");
@@ -57,7 +60,7 @@ module.exports = function(grunt) {
         dbDump(target_options, target_backup_paths);
 
         // Import dump to target DB
-        db_import(target_options, local_backup_paths.file);
+        dbImport(target_options, local_backup_paths.file);
 
         grunt.log.subhead("Operations completed");
     });
@@ -86,8 +89,8 @@ module.exports = function(grunt) {
         var local_options       = grunt.config.get('deployments').local;
 
         // Generate required backup directories and paths
-        var local_backup_paths  = generate_backup_paths("local", task_options);
-        var target_backup_paths = generate_backup_paths(target, task_options);
+        var local_backup_paths  = generateBackupPaths("local", task_options);
+        var target_backup_paths = generateBackupPaths(target, task_options);
 
         // Start execution
         grunt.log.subhead("Pulling database from '" + target_options.title + "' into Local");
@@ -101,93 +104,11 @@ module.exports = function(grunt) {
         dbDump(local_options, local_backup_paths);
 
         // Import dump into Local
-        db_import(local_options,target_backup_paths.file);
+        dbImport(local_options,target_backup_paths.file);
 
         grunt.log.subhead("Operations completed");
 
     });
-
-
-
-    function generate_backup_paths(target, task_options) {
-
-        var rtn = [];
-
-        var backups_dir = task_options['backups_dir'] || "backups";
-
-        // Create suitable backup directory paths
-        rtn['dir'] = grunt.template.process(tpls.backup_path, {
-            data: {
-                backups_dir: backups_dir,
-                env: target,
-                date: grunt.template.today('yyyymmdd'),
-                time: grunt.template.today('HH-MM-ss'),
-            }
-        });
-
-
-        rtn['file'] = rtn['dir'] + '/db_backup.sql';
-
-        return rtn;
-    }
-
-
-    /**
-     * Imports a .sql file into the DB provided
-     */
-    function db_import(config, src) {
-
-        var cmd;
-
-        // 1) Create cmd string from Lo-Dash template
-        var tpl_mysql = grunt.template.process(tpls.mysql, {
-            data: {
-                host: config.host,
-                user: config.user,
-                pass: config.pass,
-                database: config.database,
-                path: src,
-                port: config.port || 3306
-            }
-        });
-
-
-        // 2) Test whether target MYSQL DB is local or whether requires remote access via SSH
-        if (typeof config.ssh_host === "undefined") { // it's a local connection
-            grunt.log.writeln("Importing into local database");
-            cmd = tpl_mysql + " < " + src;
-        } else { // it's a remote connection
-            var tpl_ssh = grunt.template.process(tpls.ssh, {
-                data: {
-                    host: config.ssh_host
-                }
-            });
-
-            grunt.log.writeln("Importing DUMP into remote database");
-
-            cmd = tpl_ssh + " '" + tpl_mysql + "' < " + src;
-        }
-
-         // Execute cmd
-        shell.exec(cmd);
-
-        grunt.log.oklns("Database imported succesfully");
-    }
-
-
-
-    
-
-
-    
-
-
-
-
-
-
-
-
 };
